feat: typescript support to socialMedia files
